requirements.txt
absl-py==2.3.1
addict==2.4.0
basicsr==1.4.2
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
facexlib==0.3.0
filelock==3.20.0
filterpy==1.4.5
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.76.0
idna==3.11
imageio==2.37.0
intel-openmp==2021.4.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
lmdb==1.7.5
MaaAgentBinary==1.0.1
MaaFw==4.5.6
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mkl==2021.4.0
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
polars==1.34.0
polars-runtime-32==1.34.0
protobuf==6.33.0
psutil==7.1.0
pyparsing==3.2.5
pypinyin==0.55.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
realesrgan==0.3.0
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
StrEnum==0.4.15
sympy==1.14.0
tb-nightly==2.21.0a20251023
tbb==2021.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.10.16
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
ultralytics==8.3.213
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
yapf==0.43.0
